Be compatible with being invoked as pre-commit hook
git pre-commit hooks are invoked with `GIT_DIR` set (to a non-absolute
path) in environ; having that set at all is incompatible with looking to
see if a _different_ directory, via `git -C`, is a git setup.  We got
lucky that `.git` was not qualified, so we got an error message for it
being missing (when I stopped discarding stderr), rather than it always
succeeding incorrectly.