Let RUNTIME_SUPGIDS provide default group names too
Before, RUNTIME_SUPGIDS was a whitespace separated list of numeric GIDs.
After, each item in the whitespace separated list is either a numeric
GID `<nnn>` or a colon-pair of a GID and a default group name,
`<nnn>:<groupname>`.

This control will not rename any groups which already exist, but will
let a user be added to a numeric group not already in the system groups
database.  I use this with a variant build run locally.

Alpine has getent(1), so switch to that from /etc/groups hard-coding.
(I hope this is a change in Alpine and not a sign that I was just too
lazy before.)