retry_n_run now wraps retry_whileexit_e_n_run
I want the ability to only retry a command it if exits with a particular status
code; eg, if it prompted for a passphrase and the phrase was bad and it uses a
particular return-code for that.

retry_n_run is now a convenience wrapper for retry_whileexit_e_n_run.

In particular, it's now possible to retry a command while it returns true.