retry_whileexit_e_n_run: restore warn_count if success
The `retry_whileexit_e_n_run()` behind `retry_n_run()` uses `warn()` and so
increments `$warn_count`.

In practice, this is used to retry for passphrase prompting, because that's
where errors creep in such that retrying is likely to be beneficial.  I don't
think that marking a step, or a program, as bad for encountering warnings
because a passphrase was entered incorrectly once is the best approach.

So now, iff we're exiting success then restore warn_count to what it was on
entry.

Note that in the start_of_day.sh user, all but one instances were inside
command substitution sub-shells and so the warn_count wasn't actually mutated.
Which left the one place where the count did go up, and mark a step as red, was
an anomaly.  So in practice this change is restoring consistency, even while it
masks that a warning count bump in a subshell gets dropped.