stderr color: if color & emoji, use \e[G to align
Emojis can be shown with various display widths and those can change based on
the font, terminal, versions of things, and more.  So alignment is tricky.

But if we are in a mode where we are emitting emoji and color escape sequences,
then just emit an escape sequence to move to column 4.  That's either going to
be a no-op or an alignment fix.

Don't do this for non-color modes, which I'm treating as "no escape sequences
mode".